Some touchup
- Deleted sound files that we are not using
- Fixed a bug where upgrade option appears during the wave
- Wave and life is now displayed on top of the screen
- When placing a turret you will hear a cute “bleep” sound effect

ps. Game Scene scene file is not synced here, you might have slight
issues but it’s just all about assigning prefabs/materials/etc to the
script in the inspector view, so once we are done with changing stuff
on the Game Scene scene file this can be fixed as well.